backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pandas>=2.1.0
databricks-langchain>=0.1.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.49.0
pydantic>=2.0.0
python-multipart>=0.0.6
